准备生成第一个安装包

+++ build.gradle.kts
@@ -12,18 +12,29 @@ version = "1.0-SNAPSHOT"
 tasks.register("buildGuidePdf") {
     val mdFile = project.file("docs/guide/Guide.md")
     val generatedResources = project.file("build/generated/resources")
-    generatedResources.deleteRecursively()
-    generatedResources.mkdirs()
-    val outputFile = File(generatedResources, "Guide.pdf")
+    val guideDir = File(generatedResources, "guide")
+    val outputFile = File(guideDir, "Guide.pdf")
 
-    val command = arrayOf("pandoc", mdFile.absolutePath, "-o" , outputFile.absolutePath, "--pdf-engine=xelatex", "-V", "CJKmainfont=KaiTi")
-    println(command.joinToString(" "))
-    val process = ProcessBuilder(*command)
-        .start()
-    process.waitFor()
-    process.inputStream.readAllBytes().decodeToString().let(::println)
-    process.errorStream.readAllBytes().decodeToString().let(::println)
-    println("finish")
+    doLast {
+        if (mdFile.exists().not()) {
+            logger.log(LogLevel.WARN, "mdFile not exist, please create $mdFile.")
+        } else if (outputFile.exists() && outputFile.lastModified() > mdFile.lastModified()) {
+            logger.log(LogLevel.INFO, "Guide.pdf is UPDATE TO DATE.")
+        } else {
+            logger.log(LogLevel.INFO, "Guide.pdf not found, generating it.")
+            generatedResources.deleteRecursively()
+            generatedResources.mkdirs()
+            guideDir.mkdirs()
+            val command = arrayOf("pandoc", mdFile.absolutePath, "-o" , outputFile.absolutePath, "--pdf-engine=xelatex", "-V", "CJKmainfont=KaiTi")
+            println(command.joinToString(" "))
+            val process = ProcessBuilder(*command)
+                .directory(mdFile.parentFile)
+                .start()
+            process.waitFor()
+            process.inputStream.readAllBytes().decodeToString().let(::println)
+            process.errorStream.readAllBytes().decodeToString().let(::println)
+        }
+    }
 }
 
 tasks.withType<ProcessResources> {
@@ -66,14 +77,14 @@ compose.desktop {
 
         nativeDistributions {
             targetFormats(TargetFormat.Msi, TargetFormat.Exe)
-            packageVersion = "1.3.2"
+            packageVersion = "1.4.0"
             description = "A simple Windows desktop workspace manager."
 
             windows {
                 packageName = "top_ntutn_ZDesktopManager"
                 dirChooser = true
                 menuGroup = "ntutn"
-                upgradeUuid = "805a7041-2df5-40b5-96a4-2302b356ea45" // generate with https://www.guidgen.com
+                upgradeUuid = "a2fde22b-ec5f-49e4-b63f-a4235ef4848c" // generate with https://www.guidgen.com
                 // licenseFile.set(project.file("LICENSE.txt")) // 不能显示中文
             }
         }
